code_challenge_django/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0.3
djangorestframework==3.14.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
sqlparse==0.4.4
typing_extensions==4.10.0
